+++ philo/src/time.c
@@ -6,13 +6,13 @@
 /*   By: tmina-ni <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/08 16:56:19 by tmina-ni          #+#    #+#             */
-/*   Updated: 2024/02/10 01:03:46 by tmina-ni         ###   ########.fr       */
+/*   Updated: 2024/04/24 17:12:13 by tmina-ni         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../include/philo.h"
 
-unsigned long	get_current_time_ms(void)
+int	get_current_time_ms(void)
 {
 	struct timeval	tv;
 	int	time_ms;
@@ -22,26 +22,26 @@ unsigned long	get_current_time_ms(void)
 	return (time_ms);
 }
 
-unsigned long	calc_elapsed_ms(unsigned long start_time_ms)
+int	calc_elapsed_ms(int start_time_ms)
 {
-	unsigned long	elapsed_time_ms;
+	int	elapsed_time_ms;
 
 	elapsed_time_ms = get_current_time_ms() - start_time_ms;
 	return (elapsed_time_ms);
 }
 
-unsigned long	calc_elapsed_usec(unsigned long start_time_ms)
+int	calc_elapsed_usec(int start_time_ms)
 {
-	unsigned long	elapsed_time_ms;
+	int	elapsed_time_ms;
 
 	elapsed_time_ms = get_current_time_ms() - start_time_ms;
 	return (elapsed_time_ms * 1000);
 }
 
-void	ft_usleep(unsigned long usec_sleep_time)
+void	ft_usleep(int usec_sleep_time)
 {
-	unsigned long	start_time_ms;
-	unsigned long	usec_left;
+	int	start_time_ms;
+	int	usec_left;
 
 	start_time_ms = get_current_time_ms();
 	while ((calc_elapsed_usec(start_time_ms) < usec_sleep_time))
